// server/src/clients/libtorrent/client/models/Torrent.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Torrent = {
    name: string;
    info_hash: string;
    download_speed: number;
    upload_speed: number;
    downloaded: number;
    uploaded: number;
    progress: number;
    total: number;
};